Refactor context policy into SRP modules
